fix!: now UUID based on the dirigera `device.id`.
BREAKING CHANGE: Before it was generated based on a tuple { hubId, deviceType, serialNumber} which worked fine for all IKEA devices which (today) are single devices. But when connecting 3rd party devices it can fail - for example, when you have a in-wall dual relay, each relay is considered a separate dirigera device which sharing a single physical device (hence also the same serial number). In such scenario, the only thing that differentiates these two devices is their associated dirigera id. This change will cause the existing registered accessories to deregister and register again under the new UUID (which will effectively cause the devices to be removed and re-added to homekit under the default room).